chore: rename pg_replicate to supabase_etl
