+++ api/tests/common/database.rs
@@ -8,7 +8,7 @@ use sqlx::PgPool;
 /// from the "./migrations" directory after creation. Returns a [`PgPool`]
 /// connected to the newly created and migrated database. Panics if database
 /// creation or migration fails.
-pub async fn create_pg_replicate_api_database(options: &PgDatabaseOptions) -> PgPool {
+pub async fn create_supabase_etl_api_database(options: &PgDatabaseOptions) -> PgPool {
     let connection_pool = create_pg_database(options).await;
 
     sqlx::migrate!("./migrations")
