+++ api/tests/common/mod.rs
@@ -1,4 +1,4 @@
-//! Common test utilities for pg_replicate API tests.
+//! Common test utilities for supabase_etl API tests.
 //!
 //! This module provides shared functionality used across integration tests:
 //!
