+++ supabase_etl/src/conversions/text.rs
