+++ supabase_etl/src/pipeline/batching/data_pipeline.rs
