+++ supabase_etl/src/pipeline/batching/stream.rs
