+++ supabase_etl/src/pipeline/destinations/duckdb/destination.rs
