+++ supabase_etl/src/pipeline/destinations/duckdb/executor.rs
@@ -166,7 +166,7 @@ impl DuckDbExecutor {
 
     fn get_resumption_state(&self) -> Result<PipelineResumptionState, DuckDbExecutorError> {
         let copied_tables_table_name = TableName {
-            schema: "pg_replicate".to_string(),
+            schema: "supabase_etl".to_string(),
             name: "copied_tables".to_string(),
         };
         let copied_table_column_schemas = [ColumnSchema {
@@ -183,7 +183,7 @@ impl DuckDbExecutor {
             .create_table_if_missing(&copied_tables_table_name, &copied_table_column_schemas)?;
 
         let last_lsn_table_name = TableName {
-            schema: "pg_replicate".to_string(),
+            schema: "supabase_etl".to_string(),
             name: "last_lsn".to_string(),
         };
         let last_lsn_column_schemas = [ColumnSchema {
