+++ supabase_etl/src/pipeline/destinations/duckdb/mod.rs
