+++ supabase_etl/src/pipeline/destinations/stdout.rs
