+++ supabase_etl/tests/common/table.rs
